+++ oceantracker/event_loggers/log_polygon_entry_and_exit.py
@@ -10,7 +10,7 @@ class LogPolygonEntryAndExit(_BaseEventLogger):
     def __init__(self):
         super().__init__()
         # set up info/attributes
-        self.add_default_params({'polygon_list': PLC(None, dict, can_be_empty_list=False,
+        self.add_default_params({'polygon_list': PLC([], [dict], can_be_empty_list=False,
                                                      default_value= default_polygon_dict_params),
                                     'case_output_file_tag': PVC('inside_polygon_events',str)
                                                             })